refactor: separate detail/review, review page
