+++ src/components/client/detail/page.tsx
@@ -5,7 +5,6 @@ import { DehydratedState } from '@tanstack/react-query';
 import { HydrationBoundary } from '@tanstack/react-query';
 import { useQuery } from '@tanstack/react-query';
 
-import ReviewList from '@/components/common/review/review-list';
 import BottomFixedBar from '@/components/features/detail/bottom-fixed-bar';
 import BottomFixedBarPortal from '@/components/features/detail/bottom-fixed-bar-portal';
 import DetailNavBar from '@/components/features/detail/detail-nav-bar';
@@ -13,6 +12,7 @@ import DetailSwiper from '@/components/features/detail/detail-swiper';
 import ProductDetail from '@/components/features/detail/product-detail';
 import ProductInformation from '@/components/features/detail/product-information';
 import ProductNotice from '@/components/features/detail/product-notice';
+import ReviewPreview from '@/components/features/detail/review-preview';
 import { REVIEW_LIST } from '@/lib/constants/temp-review-list';
 import DividerIcon from '@/public/icons/item/divider.svg';
 import ReviewStarIcon from '@/public/icons/item/review-star.svg';
@@ -148,7 +148,7 @@ export default function ClientDetailPage({ itemId, itemDetailState }: ClientDeta
               <DividerIcon />
             </div>
             <div className="flex w-full flex-col items-center gap-5">
-              <ReviewList reviews={REVIEW_LIST} />
+              <ReviewPreview reviews={REVIEW_LIST} itemId={itemId} />
             </div>
           </div>
         </div>
